update UI and Script
