--- assets/Scripts/Room.meta
@@ -1,7 +1,0 @@
-{
-  "ver": "1.0.1",
-  "uuid": "7e9c5d9f-c9cb-4bc6-b166-2cdccd72fa32",
-  "isSubpackage": false,
-  "subpackageName": "",
-  "subMetas": {}
-}
