--- assets/Scripts/Room/game_controller.js.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "1.0.5",
-  "uuid": "6a9244de-37b1-41f4-b5aa-ace1591dea3c",
-  "isPlugin": false,
-  "loadPluginInWeb": true,
-  "loadPluginInNative": true,
-  "loadPluginInEditor": false,
-  "subMetas": {}
-}
