+++ assets/Sprites/Button_Rectangle_01_Convex_OffWhite.Png.meta
@@ -1,0 +1,34 @@
+{
+  "ver": "2.3.3",
+  "uuid": "c911c022-0527-4ca6-9f3c-7a5a3bd82ab6",
+  "type": "sprite",
+  "wrapMode": "clamp",
+  "filterMode": "bilinear",
+  "premultiplyAlpha": false,
+  "genMipmaps": false,
+  "packable": true,
+  "platformSettings": {},
+  "subMetas": {
+    "Button_Rectangle_01_Convex_OffWhite": {
+      "ver": "1.0.4",
+      "uuid": "aebf2333-1d29-4929-9166-32a1d8d4508e",
+      "rawTextureUuid": "c911c022-0527-4ca6-9f3c-7a5a3bd82ab6",
+      "trimType": "auto",
+      "trimThreshold": 1,
+      "rotated": false,
+      "offsetX": 0,
+      "offsetY": 0,
+      "trimX": 0,
+      "trimY": 0,
+      "width": 159,
+      "height": 104,
+      "rawWidth": 159,
+      "rawHeight": 104,
+      "borderTop": 0,
+      "borderBottom": 0,
+      "borderLeft": 0,
+      "borderRight": 0,
+      "subMetas": {}
+    }
+  }
+}
