+++ assets/Sprites/Common/Slider_Border_Tapered_03_Demo_Fill.png.meta
@@ -1,0 +1,34 @@
+{
+  "ver": "2.3.3",
+  "uuid": "f36b6932-b540-4d1c-a035-bcfd873b68fa",
+  "type": "sprite",
+  "wrapMode": "clamp",
+  "filterMode": "bilinear",
+  "premultiplyAlpha": false,
+  "genMipmaps": false,
+  "packable": true,
+  "platformSettings": {},
+  "subMetas": {
+    "Slider_Border_Tapered_03_Demo_Fill": {
+      "ver": "1.0.4",
+      "uuid": "bf93ec76-67a7-4bcc-8794-59f381a6c03a",
+      "rawTextureUuid": "f36b6932-b540-4d1c-a035-bcfd873b68fa",
+      "trimType": "auto",
+      "trimThreshold": 1,
+      "rotated": false,
+      "offsetX": 0,
+      "offsetY": 0,
+      "trimX": 0,
+      "trimY": 0,
+      "width": 4,
+      "height": 22,
+      "rawWidth": 4,
+      "rawHeight": 22,
+      "borderTop": 4,
+      "borderBottom": 4,
+      "borderLeft": 0,
+      "borderRight": 0,
+      "subMetas": {}
+    }
+  }
+}
