+++ assets/Sprites/Popup_02_White_Bg.png.meta
@@ -1,0 +1,34 @@
+{
+  "ver": "2.3.3",
+  "uuid": "29545c4f-cf73-4988-add7-3eed73a75abe",
+  "type": "sprite",
+  "wrapMode": "clamp",
+  "filterMode": "bilinear",
+  "premultiplyAlpha": false,
+  "genMipmaps": false,
+  "packable": true,
+  "platformSettings": {},
+  "subMetas": {
+    "Popup_02_White_Bg": {
+      "ver": "1.0.4",
+      "uuid": "52add117-d02e-4538-b64a-90a722789238",
+      "rawTextureUuid": "29545c4f-cf73-4988-add7-3eed73a75abe",
+      "trimType": "auto",
+      "trimThreshold": 1,
+      "rotated": false,
+      "offsetX": 0,
+      "offsetY": 0,
+      "trimX": 0,
+      "trimY": 0,
+      "width": 54,
+      "height": 52,
+      "rawWidth": 54,
+      "rawHeight": 52,
+      "borderTop": 0,
+      "borderBottom": 0,
+      "borderLeft": 0,
+      "borderRight": 0,
+      "subMetas": {}
+    }
+  }
+}
