+++ assets/Sprites/Popup_02_White_Border.png.meta
@@ -1,0 +1,34 @@
+{
+  "ver": "2.3.3",
+  "uuid": "633f0401-b047-43f2-9bbc-9b7b06277e07",
+  "type": "sprite",
+  "wrapMode": "clamp",
+  "filterMode": "bilinear",
+  "premultiplyAlpha": false,
+  "genMipmaps": false,
+  "packable": true,
+  "platformSettings": {},
+  "subMetas": {
+    "Popup_02_White_Border": {
+      "ver": "1.0.4",
+      "uuid": "c3d20db7-127c-4c4e-a76a-1c00650bdf27",
+      "rawTextureUuid": "633f0401-b047-43f2-9bbc-9b7b06277e07",
+      "trimType": "auto",
+      "trimThreshold": 1,
+      "rotated": false,
+      "offsetX": 0,
+      "offsetY": 0,
+      "trimX": 0,
+      "trimY": 0,
+      "width": 54,
+      "height": 52,
+      "rawWidth": 54,
+      "rawHeight": 52,
+      "borderTop": 0,
+      "borderBottom": 0,
+      "borderLeft": 0,
+      "borderRight": 0,
+      "subMetas": {}
+    }
+  }
+}
