+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "events": "^3.3.0"
+  }
+}
